refactor(test): split the false-green guard's case tables into their own file
false_green_skip_guard_test.go had grown to 526 lines against the project's
500-line limit. The file is new in this PR, so the violation is this change's
own rather than pre-existing debt, and it is in scope to fix here.

TestFalseGreenSkipGuardDetects and TestFalseGreenSkipGuardKnownBypasses move
verbatim into false_green_skip_guard_cases_test.go in the same package. The
detector itself, the repo-wide walk in TestNoSkipOnDatabaseSetupFailure, and
the guardedCalls and skipFuncs tables stay where they were. The split falls on
the natural seam: what the detector is stays in one file, the synthetic sources
proving what it does and does not catch move to the other. The two files are
now 329 and 205 lines.

This is a pure move. 196 lines were transplanted byte for byte, with no logic,
signature or name changes; the only authored lines are the new file's package
clause and its four-import block. Neither file carries a build tag, before or
after, which is what lets the guard keep seeing the files behind
//go:build integration that a default `go test` never compiles.

The package's test set is identical before and after: the same 5 top-level
tests and 7 subtests, all passing, none skipped, both with and without
-tags=integration. Complexity attribution is unchanged as well, with
TestNoSkipOnDatabaseSetupFailure still at 14 and TestFalseGreenSkipGuardDetects
still at 9, and both the CI and pre-commit gocyclo gates exclude _test.go.

Refs #1597
